refactor: 예약 가능 시간 조회 API URL 변경 및 관련 로직을 ReservationTime -> Theme 패키지로 이동
- 예약 가능 시간은 theme에 종속되므로, Theme 도메인 책임으로 판단하였음
