+++ src/main/java/roomescape/time/service/ReservationTimeService.java
@@ -1,22 +1,14 @@
 package roomescape.time.service;
 
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import roomescape.theme.controller.dto.ThemeResponse;
-import roomescape.theme.repository.ThemeRepository;
-import roomescape.time.controller.dto.response.AvailableReservationTimeResponse;
 import roomescape.time.controller.dto.request.CreateResrvationTimeRequest;
-import roomescape.time.controller.dto.request.GetAvailableTimesRequest;
 import roomescape.time.controller.dto.response.ReservationTimeResponse;
-import roomescape.time.controller.dto.response.ThemeReservationTimesResponse;
 import roomescape.time.domain.ReservationTime;
 import roomescape.time.repository.ReservationTimeRepository;
 import roomescape.time.repository.dto.CreateReservationTimeParams;
-import roomescape.time.repository.dto.FindReservedTimeParams;
 
 @Service
 @Transactional(readOnly = true)
@@ -24,7 +16,6 @@ import roomescape.time.repository.dto.FindReservedTimeParams;
 public class ReservationTimeService {
 
     private final ReservationTimeRepository reservationTimeRepository;
-    private final ThemeRepository themeRepository;
 
     @Transactional
     public ReservationTimeResponse addReservationTime(CreateResrvationTimeRequest request) {
@@ -43,28 +34,5 @@ public class ReservationTimeService {
     @Transactional
     public void removeRegisteredReservationTime(Long id) {
         reservationTimeRepository.deleteById(id);
-    }
-
-    public ThemeReservationTimesResponse findAllAvailableTimes(GetAvailableTimesRequest request) {
-        List<ReservationTime> reservationTimes = reservationTimeRepository.findAll();
-        FindReservedTimeParams params = new FindReservedTimeParams(request.themeId(), request.date());
-
-        Set<Long> reservedTimeIds = new HashSet<>(reservationTimeRepository.findIdByCondition(params));
-
-        List<AvailableReservationTimeResponse> availableTimes = reservationTimes.stream()
-                .map(time -> createResponse(time, reservedTimeIds))
-                .filter(response -> isMatchCondition(request.available(), response.available()))
-                .toList();
-        ThemeResponse theme = ThemeResponse.from(themeRepository.findById(request.themeId()));
-        return ThemeReservationTimesResponse.from(theme, availableTimes);
-    }
-
-    private AvailableReservationTimeResponse createResponse(ReservationTime time, Set<Long> reservedIdSet) {
-        boolean isAvailable = !reservedIdSet.contains(time.getId());
-        return new AvailableReservationTimeResponse(time.getId(), time.getStartAt(), isAvailable);
-    }
-
-    private boolean isMatchCondition(Boolean requestAvailable, boolean isAvailable) {
-        return requestAvailable == null || requestAvailable.equals(isAvailable);
     }
 }
