+++ src/test/java/roomescape/api/ReservationTimeApiTest.java
@@ -1,11 +1,9 @@
 package roomescape.api;
 
-import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
-import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.HashMap;
 import java.util.Map;
@@ -13,8 +11,6 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.theme.domain.Theme;
-import roomescape.time.domain.ReservationTime;
 import roomescape.util.TestDataInitializer;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
@@ -72,32 +68,4 @@ class ReservationTimeApiTest {
                 .statusCode(200)
                 .body("size()", is(1));
     }
-
-    @Test
-    void 예약_가능_시간_조회() {
-        // given
-        ReservationTime ten = dataInitializer.initializeReservationTime(LocalTime.of(10, 0));
-        ReservationTime eleven = dataInitializer.initializeReservationTime(LocalTime.of(11, 0));
-        Theme theme = dataInitializer.initializeTheme("hello", "world", "/resources/image/...");
-
-        LocalDate date = LocalDate.now().plusDays(1);
-        dataInitializer.initializeReservation("라텔", date, ten.getId(), theme.getId());
-
-
-        Map<String, Object> params = new HashMap<>();
-        params.put("date", date.toString());
-        params.put("timeId", ten.getId());
-        params.put("themeId", theme.getId());
-        params.put("available", true);
-
-        // when
-        RestAssured.given().log().all()
-                .queryParams(params)
-                .when().get("/times")
-                .then().log().all()
-                .statusCode(200)
-                .body("availableTimes.size()", is(1),
-                        "availableTimes[0].startAt", equalTo(eleven.getStartAt().toString()));
-    }
-
 }
